Alias our model's name for llama.cpp
By default, llama.cpp uses the full model path as its model id. But,
that's annoying because we're just passing in the model by name and
would like to be able to reference it by name in the llama.cpp API
server. So, tell ramalama to pass the `--alias` arg into llama.cpp so
our model's id in the OpenAI API is just its name.

Signed-off-by: Ben Browning <bbrownin@redhat.com>